Fixed warning.
Rev: src/interpret.h:1.54

+++ src/interpret.h
@@ -5,7 +5,7 @@
 \*/
 
 /*
- * $Id: interpret.h,v 1.53 2000/07/30 04:33:17 hubbe Exp $
+ * $Id: interpret.h,v 1.54 2000/08/03 12:17:04 grubba Exp $
  */
 #ifndef INTERPRET_H
 #define INTERPRET_H
@@ -94,11 +94,11 @@ struct pike_frame
 
 #define pop_stack() do{ free_svalue(--Pike_sp); debug_check_stack(); }while(0)
 
-#define pop_n_elems(X)						\
+#define pop_n_elems(X)							\
  do { ptrdiff_t x_=(X); if(x_) { 					\
-   check__positive(x_,("Popping negative number of args.... (%ld) \n",(long)x_)); \
-   Pike_sp-=x_; debug_check_stack();					\
-  free_svalues(Pike_sp,x_,BIT_MIXED);				\
+   check__positive(x_, ("Popping negative number of args.... (%ld) \n", x_)); \
+   Pike_sp -= x_; debug_check_stack();					\
+   free_svalues(Pike_sp, x_, BIT_MIXED);				\
  } } while (0)
 
 #define stack_pop_n_elems_keep_top(X) \
